refactor extract functions to test utils
