+++ src/types.ts
@@ -40,6 +40,25 @@ export interface Credentials {
   secret: string;
 }
 
+export enum EnvironmentStatus {
+  CREATED = "CREATED",
+  INACTIVE = "INACTIVE",
+  ACTIVE = "ACTIVE",
+  FAILED = "FAILED",
+  TIMEOUT = "TIMEOUT",
+  WAITING_FOR_USER = "WAITING_FOR_USER",
+  DEPLOY_IN_PROGRESS = "DEPLOY_IN_PROGRESS",
+  DESTROY_IN_PROGRESS = "DESTROY_IN_PROGRESS",
+  PR_PLAN_IN_PROGRESS = "PR_PLAN_IN_PROGRESS",
+  REMOTE_PLAN_IN_PROGRESS = "REMOTE_PLAN_IN_PROGRESS",
+  DRIFT_DETECTION_IN_PROGRESS = "DRIFT_DETECTION_IN_PROGRESS",
+  TASK_IN_PROGRESS = "TASK_IN_PROGRESS",
+  ABORTING = "ABORTING",
+  ABORTED = "ABORTED",
+  NEVER_DEPLOYED = "NEVER_DEPLOYED",
+  DRIFTED = "DRIFTED",
+}
+
 export enum DeploymentStatus {
   IN_PROGRESS = "IN_PROGRESS",
   FAILURE = "FAILURE",
